requirements.txt
beautifulsoup4==4.12.2
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
faiss-cpu==1.7.4
langchain==0.0.285
langchain-community==0.0.13
langchain-text-splitters==0.0.1
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3
tqdm==4.66.1